# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SUPA-MENU APIs",
    "version": "1.0.0",
    "description": "API documentation for Your App"
  },
  "tags": [
    {
      "name": "authentication-controller",
      "description": "Operations related to authentication"
    },
    {
      "name": "restaurant-controller",
      "description": "Operations related to restaurants"
    },
    {
      "name": "menu-controller",
      "description": "Operations related to menus"
    },
    {
      "name" : "table-controller",
      "description" : "Operation related to a restaurant table"
    }   
  ],
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Save user",
        "tags": ["authentication-controller"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User saved successfully",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/auth/signin": {
      "post": {
        "summary": "Signin",
        "tags": ["authentication-controller"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "required": true
                  },
                  "password": {
                    "email" : {
                      "type" : {
                        "type" : "string",
                        "required" : true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User signin successful",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/restaurant": {
      "post": {
        "summary": "Save restaurant",
        "tags": ["restaurant-controller"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Restaurant"
                }
              }
            }
          },
        "responses": {
          "200": {
            "description": "Restaurant saved successfully",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "get": {
        "summary": "Get restaurants",
        "tags": ["restaurant-controller"],
        "responses": {
          "200": {
            "description": "List of restaurants",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "put": {
        "summary": "Update restaurants",
        "tags": ["restaurant-controller"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Restaurant"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Restaurant updated successfully",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "delete": {
        "summary": "Delete restaurants",
        "tags": ["restaurant-controller"],
        "responses": {
          "200": {
            "description": "Restaurant deleted successfully",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },

    "/restaurant/{restaurantId}/uploadFiles": {
      "post": {
        "summary": "Upload images of the resturant",
        "tags": [
          "restaurant-controller"
        ],
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "description": "ID of the restaurant to save the item",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "images" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string",
                      "format" : "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Menu item created successfully",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/restaurant/{restaurantId}/menu": {
      "post": {
        "summary": "Create a new menu item",
        "tags": [
          "menu-controller"
        ],
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "description": "ID of the restaurant to save the item",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                    "$ref": "#/components/schemas/MenuItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Menu item created successfully",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/restaurant/{restaurantId}" : {
      "get": {
        "summary": "Get restaurant",
        "tags": ["restaurant-controller"],
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "description": "ID of the restaurant to save the item",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get restaurant",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/restaurant/{menuId}/uploadFile" : {
      "post": {
        "summary": "Upload image of menu item",
        "tags": [
          "menu-controller"
        ],
        "consumes" : ["multipart/form-data"],
        "parameters": [
          {
            "name": "menuId",
            "in": "path",
            "description": "ID of the menu item",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image" : {
                      "type" : "string",
                      "format" : "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Menu item picture uploaded successfully",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/restaurant/tables": {
      "post": {
        "tags": [
          "table-controller"
        ],
        "summary": "Create a new table",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "table",
            "description": "Table object",
            "schema": {
              "type": "object",
              "properties": {
                "tableNumber": {
                  "type": "number",
                  "required": true
                },
                "restaurantId": {
                  "type": "string",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Table"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "table-controller"
        ],
        "summary": "Get all tables",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Table"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "restaurant/tables/{id}": {
      "get": {
        "tags": [
          "table-controller"
        ],
        "summary": "Get a specific table by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Table"
            }
          },
          "404": {
            "description": "Table not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "table-controller"
        ],
        "summary": "Update a table",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "table",
            "description": "Table object",
            "schema": {
              "type": "object",
              "properties": {
                "tableNumber": {
                  "type": "number",
                  "required": true
                },
                "restaurantId": {
                  "type": "string",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Table"
            }
          },
          "404": {
            "description": "Table not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "table-controller"
        ],
        "summary": "Delete a table",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Table not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "required": true
          },
          "fullName": {
            "type": "string",
            "required": true
          },
          "email": {
            "type": "string",
            "required": true
          },
          "password": {
            "type": "string",
            "required": true
          },
          "role": {
            "type": "string",
            "required": true,
            "enum": ["RESTAURANT_MANAGER", "CLIENT"]
          }
        }
      },
      "Restaurant": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "unique": true
          },
          "fullName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "ownerName": {
            "type": "string"
          },
          "ownerPhoneNumber": {
            "type": "string"
          },
          "ownerEmail": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": ["restaurant" , "pub", "hotel", "coffeeshop", "other"]
          },
          "description": {
            "type": "string"
          },
          "openingHours": {
            "type": "string"
          },
          "closingHours": {
            "type": "string"
          } 
        },
        "required" : ["name", "fullName", "phoneNumber", "ownerName", "ownerEmail", "category", "description", "openingHours", "closingHours"]
      },
      "MenuItem": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "unique": true
          },
          "price": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": ["Drink", "Starter", "Appetizer", "Dessert", "Main"]
          }
        }
      },
      "Table": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "tableNumber": {
            "type": "number"
          },
          "restaurantId": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }, 
  "security" : [
    {
      "bearerAuth" : []
    }
  ]
}